Big update

+++ BeeFit.API/Dtos/DishesIngredients/DishesIngredientForGetDto.cs
@@ -5,6 +5,6 @@ namespace BeeFit.API.Dtos.DishesIngredients
     public class DishesIngredientForGetDto
     {
         public virtual IngredientForGetDto Ingredient { get; set; }
-        public double Quantity { get; set; }
+        public float Quantity { get; set; }
     }
 }
